// src/store/types/TDropdownSlice.tsx
import { TDropdown } from "./Common/TDropdown";

export type TDropdownSlice = {
  GENDER: TDropdown[];
  AGE: TDropdown[];
  HEIGHT_INCH: TDropdown[];
  HEIGHT_FEET: TDropdown[];
  POI: TDropdown[];
  POA: TDropdown[];
  POV: TDropdown[];
  CUSTOMER_TYPE: TDropdown[];
  CITY: TDropdown[];
  AREA_TOWN: TDropdown[];
  OCCUPATION: TDropdown[];
  INCOME_SOURCE: TDropdown[];
  FAMILY_RELATIONSHIP: TDropdown[];
  ANNUAL_INCOME: TDropdown[];
  EDUCATION: TDropdown[];
  ANNUAL_INCOME_NIVA_BUPA: TDropdown[];
  NIVA_BUPA_OCCUPATION: TDropdown[];
  PERMANENT_CITY: TDropdown[];
  MAILING_CITY: TDropdown[];
  PERMANENT_AREA_TOWN: TDropdown[];
  MAILING_AREA_TOWN: TDropdown[];
};